Out of order OpenCL scheduling for bigger problems

+++ include/cvortex/Particle.h
@@ -60,44 +60,44 @@ CVTX_EXPORT cvtx_Vec3f cvtx_Particle_visc_ind_dvort(
 
 CVTX_EXPORT cvtx_Vec3f cvtx_ParticleArr_ind_vel(
 	const cvtx_Particle **array_start,
-	const int num_particles,
+	const long num_particles,
 	const cvtx_Vec3f mes_point,
 	const cvtx_VortFunc *kernel);
 
 CVTX_EXPORT cvtx_Vec3f cvtx_ParticleArr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const int num_particles,
+	const long num_particles,
 	const cvtx_Particle *induced_particle,
 	const cvtx_VortFunc *kernel);
 
 CVTX_EXPORT cvtx_Vec3f cvtx_ParticleArr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const int num_particles,
+	const long num_particles,
 	const cvtx_Particle *induced_particle,
 	const cvtx_VortFunc *kernel,
 	const float kinematic_visc);
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_vel(
 	const cvtx_Particle **array_start,
-	const int num_particles,
+	const long num_particles,
 	const cvtx_Vec3f *mes_start,
-	const int num_mes,
+	const long num_mes,
 	cvtx_Vec3f *result_array,
 	const cvtx_VortFunc *kernel);
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const int num_particles,
+	const long num_particles,
 	const cvtx_Particle **induced_start,
-	const int num_induced,
+	const long num_induced,
 	cvtx_Vec3f *result_array,
 	const cvtx_VortFunc *kernel);
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const int num_particles,
+	const long num_particles,
 	const cvtx_Particle **induced_start,
-	const int num_induced,
+	const long num_induced,
 	cvtx_Vec3f *result_array,
 	const cvtx_VortFunc *kernel,
 	const float kinematic_visc);
